fix: move class to new package
