+++ asyncapi-core/src/main/java/com/asyncapi/v2/_0_0/model/schema/Schema.java
@@ -1,6 +1,6 @@
 package com.asyncapi.v2._0_0.model.schema;
 
-import com.asyncapi.v2.jackson.SchemasAdditionalPropertiesDeserializer;
+import com.asyncapi.v2._0_0.jackson.model.schema.SchemasAdditionalPropertiesDeserializer;
 import com.asyncapi.v2._0_0.model.ExternalDocumentation;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
